Updated metadata fields of stored knowledge records
Previously we used the metadata fields `source`, `url` and `filename`, `modified_at` which were related to files stored in an online repository (online drive, etc.)

The new fields are `source_name` and `source_id` and `modified_at`. The reason is that these fields are more generic and can be used with other sources as well (e.g. internal forum posts, wikis, etc.)