Added option to use ChromaDB instead of Milvus.
The idea is to provide alternatives for switching the vector DB with a different vector DB, as needed. The default vector DB is still Milvus, but if needed, the user can relatively easily switch to ChromaDB.

Details:

* Created a new base class VectorDB that serves as a base interface for all future supported vector DBs.
* Added template ENV for ChromaDB.
* Added a `VectorDB` class that points to the correct implementation of `BaseVectorDatabase` currently in use in the system. This serves as the SSoT for which Vector DB is used in the system.
* Split `Milvus` out of the base `docker-compose` into its own `docker-compose` file and added a `ChromaDB` `docker-compose` file. This allows for easier instantiation of only the relevant containers.

TODO: Update `README.md` with the new updated `docker compose up` command and add instructions on how to switch vector DBs.