Added support for OpenAI
Added support for using OpenAI models for both embeddings and inference, instead of Bedrock models.

Bedrock and OpenAI return events with a somewhat different structure. This resulted in a reorgranization and refactoring of some of the code that handles the chat agent's events.

We've also made some changes that we've planned to do anyways, that allow for better separation of concerns between the API part and the LLM interface part.

Details

* Made sure that every point in the code interfaces the inference and embeddings models using the variables `ChatModel` and `EmbeddingsModel`. This makes it so that it'll be easier to replace their values (SSoT).
* Added necessary `.env` and `README.md` sections.
* Moved handling of events from LLM to `llm.py` and made sure that `chat.py` handles the API-related actions specifically.
* Added layers of abstraction in `llm.py` in order to make it easier to override some of the default behavior, based on the specific model used and to provide cleaner API with `ENum`s instead of "magic" values.