Indexing and CCH
Added a new method for indexing that adds context to each chunk that goes into the vector DB. This improves both the retrieval and the inference stages of the RAG system.

The reason is that this technique, called Contextual Chunk Headers, allows chunks that are in the middle of the document, to contain the overarching context, which is often missing from the chunk itself.

This also prompted a slight revamp of the splitting and indexing process in general.

Details

* Moved the indexing strategy out of the vector DB class, into a strategies submodule. This allows the user to mix and match of vector DBs and strategy classes. The indexing strategy is usually unrelated to the vector DB itself
* Added 2 indexing strategies:
  * The original one - a simple contextual split and embedding.
  * Context aware - builds on the simple one, but for each document, first asks an LLM to summarize the document and adds that summarization to each chunk.
* Added a Jupyter notebook that demonstrates the importance of CCH using a real life example.

Additional Changes

* Added a `README.md` section on how to run and connect to Jupyter Lab.
* Fixed problems with `drop_collection` in Chroma DB.
* Made `add_documents` and the indexing strategies work with iterators. That said, LangChain in some cases expects a list and not an iterator, so need to be mindful of that.
* Added payload support for the `DocumentMetadata` class
* Fixed tests following `DocumentMetadata` class introduction.